// src/utils/enums.ts
export enum VIEW_TYPE {
  DEEP_LINK_VIEW = "Deeplinkview",
  EXTENSION_WALLET_VIEW = "ExtensionWalletview",
  OKTO_WALLET_VIEW = "OktoWallet",
  TIPLINK_WALLET_VIEW = "TiplinkView",
}

export enum LOCALSTORAGE_ITEMS {
  SHARED_SECRET = "sharedSecret",
  SESSION = "session",
  WALLET_ADDRESS = "walletAddress",
  DAPP_KEY_PAIR = "dappKeyPair",
  VERIFICATION_MESSAGE = "verificationMessage",
  ACCESS_TOKEN = "accessToken",
  REFRESH_TOKEN = "refreshToken",
}

export enum DEEP_LINK_FUNCTIONS {
  SIGN_MESSAGE = "signMessage",
  CONNECT = "connect",
  DISCONNECT = "disconnect",
  SIGN_AND_SEND_TRANSACTION = "signAndSendTransaction",
}